[BUGFIX][EWS-2807] Fixed fractional part

See merge request elta/elta-android!464

+++ presentation/src/main/java/com/elta/android/presentation/features/calcutator/viewmodel/DishDetailViewModel.kt
@@ -137,7 +137,7 @@ class DishDetailViewModel @Inject constructor(
                             setText(newDish.servings.findOrFirst { saveServing.id == it.id }.servingDescription)
                         }
                     }
-                    portionCountTextField.setText(newDish.servingAmount.toInt().toString())
+                    portionCountTextField.setText(newDish.servingAmount.toString())
                 }
         }
     }
